add functions
add data-key on all divs,
create function to get the key code (getElementByKeyCode), create function to activate and add the css class activate and validate if it is success or fail to add each class.

Create a function to deactivate to reseat the class to the original class.

Setup the levels in 15, generate a function to get an array to play.